Fixed checksum upload
I found a possible bug in Artifactory Java Client. When a file is uploaded for the first time (when there is no file in the repository with the same checksum), the checksum is not set correctly. Therefore, the UI shows “Client did not publish a checksum value.” message. 

This is because the REST API returns HTTP 404 when the no file exists before. Then, the file content is uploaded, but UploadableArtifactImpl does not pass headers to uploadContent method. Therefore, the server does not save the checksum.

The patch should fix this issue.