Concurrency Support
we need the artifactory java client to support concurrent requests.

the old rest method (wrapper) would replace the underlying httpClient's
parser on runtime according to the content type and the expected response class,
this operation was not thread safe and did not meet the performance requirements that we need.

the wrapper was removed and parsing is now done explicitly by us and the HttpClient's parser is not replaced at runtime.
this makes the artifactory client thread safe in regards to rest requests.